pyproject.toml
[tool.poetry]
name = "sciloc"
version = "0.1.0"
description = ""
authors = ["Saeed Gholami <saeed.ghollami@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
beautifulsoup4 = "^4.12.2"
requests = "^2.31.0"
matplotlib = "^3.8.2"
arabic-reshaper = "^3.0.0"
python-bidi = "^0.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
